dev mode

+++ resolvers/moe.js
@@ -3,49 +3,49 @@ const msg = require("../config/messages");
 const model = require("../models/models");
 
 module.exports = {
-    Query:{
+    Query: {
 
     },
-    Mutation:{
+    Mutation: {
 
         addMOE: async (_, args, context) => {
 
-            // if (!context.req.auth.connected) {
-            //     throw new Error(msg.notConnectedUser);
-            // }
+            if (!context.req.auth.connected) {
+                throw new Error(msg.notConnectedUser);
+            }
 
 
-            let person = await model.personne.create(args.personneInfo);            
+            let person = await model.personne.create(args.personneInfo);
 
             let moe = await model.moe.create({
                 IdPersonne: person.IdPersonne,
                 moisDebut: args.moisDebut,
-                moisFin:args.moisFin,
-                salaireMens:args.salaireMens,
+                moisFin: args.moisFin,
+                salaireMens: args.salaireMens,
                 obs_moe: args.observation,
-                IdActPcpl:args.IdActivitPricip,
+                IdActPcpl: args.IdActivitPricip,
                 IdTypMOE: args.IdTypeMOE,
                 IdEA: args.IdEA,
             });
 
             // let aider = await model.aider.create({
-                // IdTypMOE: args.IdTypeMOE,
-                // IdMOE: moe.IdPersonne,
-                // IdEA: args.IdEA,
-                
+            // IdTypMOE: args.IdTypeMOE,
+            // IdMOE: moe.IdPersonne,
+            // IdEA: args.IdEA,
+
             // });
 
-            
-            let avantages = (args.avantages? args.avantages: []);
-            
+
+            let avantages = (args.avantages ? args.avantages : []);
+
             avantages.forEach(async el => {
-                await model.avantage_nat.create({   
+                await model.avantage_nat.create({
                     IdTypeAvNat: el.type,
                     puAvNat: el.puAvNat,
                     qteAvNat: el.qteAvNat,
                     IdPersonne: person.IdPersonne
                 });
-                
+
             });
 
             return {
@@ -57,18 +57,18 @@ module.exports = {
 
         updateMOE: async (_, args, context) => {
 
-            // if (!context.req.auth.connected) {
-            //     throw new Error(msg.notConnectedUser);
-            // }
+            if (!context.req.auth.connected) {
+                throw new Error(msg.notConnectedUser);
+            }
 
             console.log(args);
-            
-            model.personne.update(args.personneInfo,{
+
+            model.personne.update(args.personneInfo, {
                 where: {
                     IdPersonne: args.IdMOE
                 }
-            });   
-            
+            });
+
             let person = await model.personne.findOne({
                 where: {
                     IdPersonne: args.IdMOE
@@ -78,10 +78,10 @@ module.exports = {
 
             await model.moe.update({
                 moisDebut: args.moisDebut,
-                moisFin:args.moisFin,
-                salaireMens:args.salaireMens,
+                moisFin: args.moisFin,
+                salaireMens: args.salaireMens,
                 obs_moe: args.observation,
-                IdActPcpl:args.IdActivitPricip,
+                IdActPcpl: args.IdActivitPricip,
                 IdTypMOE: args.IdTypeMOE,
             }, {
                 where: {
@@ -89,35 +89,35 @@ module.exports = {
                 }
             });
 
-            let moe =  await model.moe.findOne({
+            let moe = await model.moe.findOne({
                 raw: true,
                 where: {
                     IdPersonne: args.IdMOE
                 },
-                
+
             });
 
             // console.log(moe);
-  
-            
-            let avantages = (args.avantages? args.avantages: []);
-            
+
+
+            let avantages = (args.avantages ? args.avantages : []);
+
             avantages.forEach(async el => {
 
-                if (el.IdAvNat){
-                    await model.avantage_nat.update({   
+                if (el.IdAvNat) {
+                    await model.avantage_nat.update({
                         IdTypeAvNat: el.type,
                         puAvNat: el.puAvNat,
                         qteAvNat: el.qteAvNat,
                         IdPersonne: person.IdPersonne
-                    },{
+                    }, {
                         where: {
                             IdAvNat: el.IdAvNat
                         }
                     });
                 }
-                else{
-                    await model.avantage_nat.create({   
+                else {
+                    await model.avantage_nat.create({
                         IdTypeAvNat: el.type,
                         puAvNat: el.puAvNat,
                         qteAvNat: el.qteAvNat,
@@ -130,22 +130,22 @@ module.exports = {
                 ...moe,
                 observation: moe.obs_moe
             }
-        
+
         },
 
         deleteMOE: async (_, args, context) => {
 
-            // if (!context.req.auth.connected) {
-            //     throw new Error(msg.notConnectedUser);
-            // }
+            if (!context.req.auth.connected) {
+                throw new Error(msg.notConnectedUser);
+            }
 
-            
+
             await model.personne.destroy({
-                where:{
+                where: {
                     IdPersonne: args.IdMOE
                 }
             });
-    
+
 
             return true;
         },
@@ -161,7 +161,7 @@ module.exports = {
         //         where: {
         //             IdPersonne: args.IdMOE
         //         },
-                
+
         //     });
 
         //     let aider = await model.aider.create({
@@ -174,9 +174,9 @@ module.exports = {
         //         obs_moe: args.observation,
         //     });
 
-            
+
         //     let avantages = (args.avantages? args.avantages: []);
-            
+
         //     avantages.forEach(async el => {
 
         //         await model.avantage_nat.create({
